updating the Cygwin build
git-svn-id: file:///desktop/SourceForge%20Backup/empyrean/svnrepos/trunk@337 fea1f14b-4e95-431d-9825-d582385f3f73

+++ empyrean/src/client/Profiler.cpp
@@ -24,7 +24,7 @@ namespace pyr {
 
     float ProfileBlock::getAverageTime() const {
         float total = 0;
-        for (int i = 0; i < REMEMBERED_FRAMES; ++i) {
+        for (size_t i = 0; i < REMEMBERED_FRAMES; ++i) {
             total += lastFrames[i].time();
         }
         return total / REMEMBERED_FRAMES;
@@ -32,7 +32,7 @@ namespace pyr {
 
     float ProfileBlock::getAverageTotalTime() const {
         float total = 0;
-        for (int i = 0; i < REMEMBERED_FRAMES; ++i) {
+        for (size_t i = 0; i < REMEMBERED_FRAMES; ++i) {
             total += lastFrames[i].timePlusChildren;
         }
         return total / REMEMBERED_FRAMES;
